html/semantics/scripting-1/the-script-element/module/custom-element-exception.html

<!DOCTYPE html>
<title>Handling of exceptions in custom element constructors</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    const test_load = async_test(
        "Test that exceptions from the constructor of a custom element " +
        "inside a module are propagated as expected.\n");
    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_true(window.exception instanceof Error);
      assert_equals(exception.message, "custom element error");
    }));
</script>
<script type="module">
    class XThrower extends HTMLElement {
      constructor() {
        super();
        throw new Error("custom element error");
      }
    }
    customElements.define("x-thrower", XThrower);
    try {
      document.createElement("x-thrower");
    } catch (e) {
      window.exception = e;
    }
</script>
