LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/custom-element-exception-expected.txt

CONSOLE MESSAGE: Error: custom element error

PASS Test that exceptions from the constructor of a custom element inside a module are propagated as expected.
